增加连接页面

+++ src/main/server/router.ts
@@ -2,13 +2,13 @@
  * @Author: JOY
  * @Date: 2024-06-21 10:57:35
  * @LastEditors: JOY
- * @LastEditTime: 2024-06-21 16:41:06
+ * @LastEditTime: 2024-06-24 15:20:11
  * @Description:
  */
 import express from "express";
 import fs from "node:fs";
 import { ROOT_REDIS_CONF } from "./constants";
-import { GroupDto } from "./dto";
+import { GroupAddDto, GroupEditDto, GroupDelDto } from "./dto";
 import { success, fail } from "./utils";
 
 import type { Link } from "./server.d";
@@ -28,7 +28,7 @@ router.get("/", (_, res) => {
 let CacheRedis: Link[] = [];
 // 新增分组
 router.post("/sys/group/add", (req, res) => {
-  const groupDto = req.body as GroupDto;
+  const groupDto = req.body as GroupAddDto;
   // 获取组名称
   const groupName = groupDto.group_name;
 
@@ -39,15 +39,15 @@ router.post("/sys/group/add", (req, res) => {
     const groups = (JSON.parse(read) || []) as Link[];
     // 判断数据是否存在
     const group = groups.find(
-      (item) => item.pid === parseInt(groupDto.group_id) && item.name === groupName
+      (item) => item.pid === parseInt(groupDto.group_pid) && item.name === groupName
     );
     if (group) return fail(1001, "当前组已存在", res);
 
     const ngroup: Link = {
       id: new Date().getTime(),
-      pid: parseInt(groupDto.group_id),
+      pid: parseInt(groupDto.group_pid),
       name: groupName,
-      type: '1',
+      type: "1",
       host: "",
       user: "",
       pwd: "",
@@ -60,7 +60,7 @@ router.post("/sys/group/add", (req, res) => {
       id: new Date().getTime(),
       pid: 0,
       name: groupName,
-      type: '1',
+      type: "1",
       host: "",
       user: "",
       pwd: "",
@@ -68,6 +68,57 @@ router.post("/sys/group/add", (req, res) => {
     };
     const groups = [group];
     CacheRedis = groups;
+  }
+  // 写入的内容是JSON
+  fs.writeFileSync(ROOT_REDIS_CONF, JSON.stringify(CacheRedis));
+  return success(CacheRedis, res);
+});
+
+// 修改分组
+router.post("/sys/group/edit", (req, res) => {
+  const groupDto = req.body as GroupEditDto;
+  // 读取文件
+  if (fs.existsSync(ROOT_REDIS_CONF)) {
+    // 读取本地配置
+    const read = fs.readFileSync(ROOT_REDIS_CONF, { encoding: "utf-8" });
+    // 解析读取的数据
+    const groups = (JSON.parse(read) || []) as Link[];
+    // 判断数据是否存在
+    const group = groups.find((item) => item.id === parseInt(groupDto.group_id));
+    if (!group) return fail(1002, "当前组不存在", res);
+
+    // 当前组名是否存在
+    const nameGroup = groups.find(
+      (item) => item.pid === parseInt(groupDto.group_pid) && item.name === groupDto.group_name
+    );
+    if (nameGroup) return fail(1001, "当前组已存在", res);
+
+    group.pid = +groupDto.group_pid;
+    group.name = groupDto.group_name;
+
+    CacheRedis = groups;
+  }
+  // 写入的内容是JSON
+  fs.writeFileSync(ROOT_REDIS_CONF, JSON.stringify(CacheRedis));
+  return success(CacheRedis, res);
+});
+
+// 删除分组
+router.delete("/sys/group/del", (req, res) => {
+  const groupDto = req.body as GroupDelDto;
+  // 读取文件
+  if (fs.existsSync(ROOT_REDIS_CONF)) {
+    // 读取本地配置
+    const read = fs.readFileSync(ROOT_REDIS_CONF, { encoding: "utf-8" });
+    // 解析读取的数据
+    const groups = (JSON.parse(read) || []) as Link[];
+    // 判断数据是否存在
+    const index = groups.findIndex((item) => item.id === parseInt(groupDto.group_id));
+    if (index < 0) return fail(1002, "当前组不存在", res);
+    // 注意，删除组的时候，如果改组存在子组也需要进行删除
+    CacheRedis = groups.filter(
+      (item) => item.id !== parseInt(groupDto.group_id) && item.pid !== parseInt(groupDto.group_id)
+    );
   }
   // 写入的内容是JSON
   fs.writeFileSync(ROOT_REDIS_CONF, JSON.stringify(CacheRedis));
@@ -85,4 +136,5 @@ router.get("/sys/links", (_, res) => {
   }
   return success(CacheRedis, res);
 });
+
 export default router;
